feat(Model): Allow customizing the name of generated schemas. (#2542)
## Description

Adds a new `name` property the `#[Model]` attribute & `Model` class for
registering schema's.

This makes it possible to customize the name that needs to be registered
for a model schema.

This PR also comes with a refactor to the "alternative names" logic, to
ensure deduplication logic (introduced in
https://github.com/nelmio/NelmioApiDocBundle/pull/2461) is also retained
for manually registered schema's through the bundle config.

> [!NOTE]
> This PR, also changes that manually registered schema's get registered
**first** now. (Instead of registering these last)

Closes #2540, closes #2113, closes #2515

## What type of PR is this? (check all applicable)
- [ ] Bug Fix
- [x] Feature
- [x] Refactor
- [ ] Deprecation
- [ ] Breaking Change
- [ ] Documentation Update
- [ ] CI

## Checklist
- [x] I have made corresponding changes to the documentation (`docs/`)
- [ ] I have made corresponding changes to the changelog
(`CHANGELOG.md`)