feat: Implement Ulid type description in ClassDescriber (#2556)
Add support for Ulid type in ClassDescriber.

As the schema format is not bounded to any list, we use `ulid` for Ulid
types here.

## Description

This allows Ulid properties to be properly labbelled as `type: string,
format: ulid` in the schema definition.

In turns, this allows data to be passed using one of the (short) forms
of ULIDs too

## What type of PR is this? (check all applicable)
- [x] Bug Fix
- [x] Feature
- [ ] Refactor
- [ ] Deprecation
- [ ] Breaking Change
- [ ] Documentation Update
- [ ] CI

## Checklist
- [ ] I have made corresponding changes to the documentation (`docs/`)
- [ ] I have made corresponding changes to the changelog
(`CHANGELOG.md`)